+++ src/Validation/Validation.ts
@@ -1157,7 +1157,7 @@ export class Validation {
     }
 
     /**
-     * To check if the field exist if its peers exists too.
+     * To check if the field exist if any of its peers exists too.
      *
      * @param validation_options
      */
@@ -1208,6 +1208,67 @@ export class Validation {
                 const doesCurrentFieldExist = Boolean(value);
 
                 if (anyOfTheOtherFieldExists && !doesCurrentFieldExist) {
+                    return Promise.reject(message);
+                }
+
+                return Promise.resolve();
+            })
+
+        }
+    }
+
+    /**
+     * To check if the field exist if all of its peers exists too.
+     *
+     * @param validation_options
+     */
+    static requiredWithAll(validation_options: IValidationRequiredWithDto) : Function {
+
+        const {
+            checkIn = "any",
+            params = {
+                fields: []
+            }
+        } = validation_options;
+
+        let {
+            message = `The :attribute required with all of the following fields ${params.fields}`
+        } = validation_options;
+
+        if (!Boolean(params.fields)) {
+            throw new Error(`(requiredWithAll) validation, params.fields must not be empty array`);
+        }
+
+        if (!Array.isArray(params.fields)) {
+            throw new Error(`(requiredWithAll) validation, params.fields must be of type array`);
+        }
+
+        return (field : string) => {
+
+            const toMatch = Util.returnBasedOnCheckIn(checkIn,field);
+
+            message = Util.replaceMessageWithField(field, message);
+
+            return toMatch.custom((value, { req, location }) => {
+
+                const requestObject =
+                    location === "body"
+                        ? req.body
+                        : location === "query"
+                            ? req.query
+                            : location === "headers"
+                                ? req.headers
+                                : req.params;
+
+                const allOfTheFieldsExists = params.fields.every((otherField : string) =>
+                    Boolean(
+                        get(requestObject, otherField, false)
+                    )
+                );
+
+                const doesCurrentFieldExist = Boolean(value);
+
+                if (allOfTheFieldsExists && !doesCurrentFieldExist) {
                     return Promise.reject(message);
                 }